refactor(tests): regroup test files
